add matching logic and cleanup
