+++ docs/api/sample_id.rst
@@ -19,9 +19,6 @@ Submodules
    sample_id.analysis
    sample_id.audio
    sample_id.cli
-   sample_id.deprecated_ann
-   sample_id.deprecated_main
-   sample_id.hough
    sample_id.io
    sample_id.main
    sample_id.trackio
