+++ sample_id/ann/ann.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 
 import abc
+import itertools
 import logging
 import os
 import tempfile
 import zipfile
-from typing import Iterable, Optional
+from typing import Any, Iterable, List, Optional, Sequence
 
 import numpy as np
 
@@ -43,7 +44,7 @@ class Matcher(abc.ABC):
         pass
 
     @abc.abstractmethod
-    def nearest_neighbors(self, fp: Fingerprint, k: int = 1) -> str:
+    def nearest_neighbors(self, fp: Fingerprint, k: int = 1) -> Iterable[Match]:
         """Fetch nearest neighbors to this fingerprint's keypoints."""
         pass
 
@@ -67,7 +68,7 @@ class Matcher(abc.ABC):
             self.add_fingerprint(fingerprint, **kwargs)
         return self
 
-    def can_add_fingerprint(self, fingerprint: Fingerprint) -> Boolean:
+    def can_add_fingerprint(self, fingerprint: Fingerprint) -> bool:
         """Check if fingerprint can be added to matcher."""
         if not self.meta.sr:
             self.meta.sr = fingerprint.sr
@@ -87,7 +88,7 @@ class Matcher(abc.ABC):
             tmp_model_path = os.path.join(tmpdir, MATCHER_FILENAME)
             tmp_meta_path = os.path.join(tmpdir, META_FILENAME)
             tmp_model_path = self.save_model(tmp_model_path)
-            self.meta.save(tmp_meta_path)
+            self.meta.save(tmp_meta_path, compress=compress)
             with zipfile.ZipFile(filepath, "w", compression=zipfile.ZIP_DEFLATED) as zipf:
                 logger.info(f"Zipping {tmp_model_path} and {tmp_meta_path} into {zipf.filename}")
                 zipf.write(tmp_model_path, arcname=MATCHER_FILENAME)
@@ -108,14 +109,14 @@ class Matcher(abc.ABC):
     @classmethod
     def from_fingerprint(cls, fp: Fingerprint, **kwargs) -> Matcher:
         """Useful for determining metadata for the Matcher based on the data being added."""
-        matcher = cls.create(fp.descriptors.shape[1], sr=fp.sr, hop_length=fp.hop_length, **kwargs)
+        matcher = cls.create(sr=fp.sr, hop_length=fp.hop_length, n_features=fp.descriptors.shape[1], **kwargs)
         return matcher.add_fingerprint(fp, **kwargs)
 
     @classmethod
-    def from_fingerprints(cls, fingerprints: Iterable[Fingerprint], **kwargs) -> Matcher:
+    def from_fingerprints(cls, fingerprints: Sequence[Fingerprint], **kwargs) -> Matcher:
         """My data is small, just create and train the entire matcher."""
         fp = fingerprints[0]
-        matcher = cls.create(fp.descriptors.shape[1], sr=fp.sr, hop_length=fp.hop_length, **kwargs)
+        matcher = cls.create(sr=fp.sr, hop_length=fp.hop_length, n_features=fp.descriptors.shape[1], **kwargs)
         return matcher.add_fingerprints(fingerprints, **kwargs)
 
     @classmethod
@@ -200,12 +201,179 @@ class MatcherMetadata:
 
 
 class Match(object):
-    def __init__(self, keypoint, neighbors: Iterable[Neighbor]):
-        self.keypoint = keypoint
+    def __init__(self, keypoint, neighbors: Sequence[Neighbor]):
+        self.keypoint = Keypoint(keypoint)
         self.neighbors = neighbors
 
 
 class Neighbor(object):
-    def __init__(self, index: int, distance: float):
+    def __init__(self, index: int, distance: float, meta: MatcherMetadata):
         self.index = index
         self.distance = distance
+        self.keypoint = Keypoint(meta.index_to_kp[index])
+        self.source_id = meta.index_to_id[index]
+
+
+class Keypoint(object):
+    def __init__(self, kp):
+        self.x = kp[0]
+        self.y = kp[1]
+        self.scale = kp[2]
+        self.orientation = kp[3]
+        self.kp = kp
+
+
+def filter_matches(
+    matches: List[Match],
+    abs_thresh=None,
+    ratio_thresh=None,
+    cluster_dist=20,
+    cluster_size=1,
+    match_orientation=True,
+    ordered=False,
+):
+    logger.info("Filtering nearest neighbors down to actual matched samples")
+    if match_orientation:
+        # Remove matches with differing orientations
+        total = len(matches)
+        for match in matches:
+            orient = match.keypoint.orientation
+            while match.neighbors and abs(orient - match.neighbors[0].keypoint.orientation) > 0.2:
+                match.neighbors = match.neighbors[1:]
+            if not match.neighbors == 0:
+                matches.remove(match)
+            # elif len(match.neighbors) < 2:
+            #     # logger.warn('Orientation check left < 2 neighbors')
+            #     matches.remove(match)
+        logger.info("Differing orientations removed: {}, remaining: {}".format(total - len(matches), len(matches)))
+    if abs_thresh:
+        # Apply absolute threshold
+        total = len(matches)
+        matches = [match for match in matches if match.neighbors[0].distance < abs_thresh]
+        logger.info("Absolute threshold removed: {}, remaining: {}".format(total - len(matches), len(matches)))
+    if ratio_thresh:
+        # Apply ratio test
+        total = len(matches)
+        for match in list(matches):
+            n1 = match.neighbors[0]
+            n2 = next((n for n in match.neighbors if n.source_id != n1.source_id), None)
+            if n2 is None:
+                logger.warn("No second neighbor for ratio test, consider increasing k")
+                d2 = n1.distance * 2
+            else:
+                d2 = n2.distance
+            if not (n1.distance < ratio_thresh * d2):
+                matches.remove(match)
+        logger.info("Ratio threshold removed: {}, remaining: {}".format(total - len(matches), len(matches)))
+    # TODO: replace this hackier clustering
+    # Only keep when there are multiple within a time cluster
+    clusters = list(cluster_matches(matches, cluster_dist))
+    filtered_clusters = [
+       cluster for cluster in clusters if len(cluster) >= cluster_size
+    ]
+    # TODO: do hough clustering instead
+    # filtered_clusters, clusters = hough.cluster(matches, cluster_size, cluster_dist)
+    # logger.info("Total Clusters: {}, filtered clusters: {}".format(len(clusters), len(filtered_clusters)))
+    if ordered:
+        orderedx_clusters = []
+        ordered_clusters = []
+        for cluster in filtered_clusters:
+            sorted_trainx = sorted(cluster, key=lambda m: m.neighbors[0].kp.x)
+            sorted_queryx = sorted(cluster, key=lambda m: m.query.x)
+            if sorted_trainx == sorted_queryx:
+                orderedx_clusters.append(cluster)
+        logger.info("Total Clusters: {}, orderedx clusters: {}".format(len(clusters), len(orderedx_clusters)))
+        for cluster in orderedx_clusters:
+            sorted_trainy = sorted(cluster, key=lambda m: m.neighbors[0].kp.y)
+            sorted_queryy = sorted(cluster, key=lambda m: m.query.y)
+            if sorted_trainy == sorted_queryy:
+                ordered_clusters.append(cluster)
+        logger.info("Total Clusters: {}, ordered clusters: {}".format(len(clusters), len(ordered_clusters)))
+        filtered_clusters = ordered_clusters
+    matches = [match for cluster in filtered_clusters for match in cluster]
+    logger.info("Filtered matches: {}".format(len(matches)))
+    return filtered_clusters
+
+
+def cluster_matches(matches, cluster_dist):
+    class Cluster(object):
+        def __init__(self, match):
+            self.min_query = match.query.x
+            self.max_query = match.query.x
+            self.min_train = match.neighbors[0].kp.x
+            self.max_train = match.neighbors[0].kp.x
+            self.matches = [match]
+
+        def add(self, match):
+            if match.query.x > self.min_query:
+                self.min_query = match.query.x
+            if match.query.x > self.max_query:
+                self.max_query = match.query.x
+            if match.neighbors[0].kp.x < self.min_train:
+                self.min_train = match.neighbors[0].kp.x
+            if match.neighbors[0].kp.x > self.max_train:
+                self.max_train = match.neighbors[0].kp.x
+            self.matches.append(match)
+
+        def merge(self, cluster):
+            if cluster.min_query < self.min_query:
+                self.min_query = cluster.min_query
+            if cluster.max_query > self.max_query:
+                self.max_query = cluster.max_query
+            if cluster.min_train < self.min_train:
+                self.min_train = cluster.min_train
+            if cluster.max_train > self.max_train:
+                self.max_train = cluster.max_train
+            self.matches.extend(cluster.matches)
+
+    logger.info("Clustering matches...")
+    logger.info("cluster_dist: {}".format(cluster_dist))
+    matches = sorted(matches, key=lambda m: (m.neighbors[0].kp.source, m.query.x))
+    clusters = {}
+    for source, group in itertools.groupby(matches, lambda m: m.neighbors[0].kp.source):
+        for match in group:
+            cluster_found = False
+            for cluster in clusters.get(source, []):
+                if (
+                    match.query.x >= cluster.min_query - cluster_dist
+                    and match.query.x <= cluster.max_query + cluster_dist
+                ) and (
+                    match.neighbors[0].kp.x >= cluster.min_train - cluster_dist
+                    and match.neighbors[0].kp.x <= cluster.max_train + cluster_dist
+                ):
+                    if not any(
+                        match.neighbors[0].kp.x == c.neighbors[0].kp.x
+                        and match.neighbors[0].kp.y == c.neighbors[0].kp.y
+                        for c in cluster.matches
+                    ):
+                        cluster_found = True
+                        cluster.add(match)
+            if not cluster_found:
+                clusters.setdefault(source, []).append(Cluster(match))
+        # Merge nearby clusters
+        merged_clusters = clusters.get(source, [])
+        for cluster in clusters.get(source, []):
+            for c in merged_clusters:
+                if (
+                    c != cluster
+                    and (
+                        cluster.min_query >= c.min_query - cluster_dist
+                        and cluster.max_query <= c.max_query + cluster_dist
+                    )
+                    and (
+                        cluster.min_train >= c.min_train - cluster_dist
+                        and cluster.max_train <= c.max_train + cluster_dist
+                    )
+                ):
+                    cluster_points = set((m.neighbors[0].kp.x, m.neighbors[0].kp.y) for m in cluster.matches)
+                    c_points = set((m.neighbors[0].kp.x, m.neighbors[0].kp.y) for m in c.matches)
+                    if cluster_points & c_points:
+                        break
+                    c.merge(cluster)
+                    logging.info(len(merged_clusters))
+                    merged_clusters.remove(cluster)
+                    logging.info(len(merged_clusters))
+                    cluster = c
+        clusters[source] = merged_clusters
+    clusters = [cluster.matches for sources in clusters.values() for cluster in sources]
+    return clusters
